feat: edit/create collection form reusable
