+++ public/locales/en/editCollectionGeneralInfo.json
@@ -1,4 +1,4 @@
 {
-  "pageTitle": "Edit Collection General Information",
+  "pageTitle": "Edit Collection",
   "notAllowedToEditCollection": "You do not have permissions to edit a collection within this collection."
 }
