+++ src/sections/Route.enum.ts
@@ -10,7 +10,7 @@ export enum Route {
   FILES = '/files',
   COLLECTIONS_BASE = '/collections',
   COLLECTIONS = '/collections/:collectionId',
-  CREATE_COLLECTION = '/collections/:ownerCollectionId/create',
+  CREATE_COLLECTION = '/collections/:parentCollectionId/create',
   EDIT_COLLECTION_GENERAL_INFO = '/collections/:collectionId/edit-general-info',
   ACCOUNT = '/account'
 }
@@ -18,7 +18,7 @@ export enum Route {
 export const RouteWithParams = {
   COLLECTIONS: (collectionId?: string) =>
     collectionId ? `/collections/${collectionId}` : Route.COLLECTIONS_BASE,
-  CREATE_COLLECTION: (ownerCollectionId: string) => `/collections/${ownerCollectionId}/create`,
+  CREATE_COLLECTION: (parentCollectionId: string) => `/collections/${parentCollectionId}/create`,
   CREATE_DATASET: (collectionId: string) => `/datasets/${collectionId}/create`,
   EDIT_COLLECTION_GENERAL_INFO: (collectionId: string) =>
     `/collections/${collectionId}/edit-general-info`
