+++ src/sections/shared/form/EditCreateCollectionForm/collection-form/metadata-fields-section/fields-from-parent-checkbox/ConfirmResetModificationsModal.tsx
@@ -1,5 +1,5 @@
-import { Alert, Button, Modal } from '@iqss/dataverse-design-system'
 import { useTranslation } from 'react-i18next'
+import { Alert, Button, Modal } from '@iqss/dataverse-design-system'
 
 interface ConfirmResetModificationsModalProps {
   showModal: boolean
