+++ src/sections/shared/form/EditCreateCollectionForm/collection-form/metadata-fields-section/metadata-input-level-fields-block/input-levels-table/InputLevelsTable.module.scss
