+++ src/sections/shared/form/EditCreateCollectionForm/collection-form/metadata-fields-section/metadata-input-level-fields-block/input-levels-table/InputLevelsTable.tsx
@@ -1,7 +1,7 @@
 import { ReactNode } from 'react'
 import cn from 'classnames'
 import { Table } from '@iqss/dataverse-design-system'
-import { MetadataBlockInfo } from '../../../../../../metadata-block-info/domain/models/MetadataBlockInfo'
+import { MetadataBlockInfo } from '@/metadata-block-info/domain/models/MetadataBlockInfo'
 import { InputLevelFieldRow } from './InputLevelFieldRow'
 import styles from './InputLevelsTable.module.scss'
