+++ src/sections/shared/form/EditCreateCollectionForm/collection-form/top-fields-section/ContactsField.tsx
@@ -2,7 +2,7 @@ import { useCallback, useMemo } from 'react'
 import { Col, Form, Row } from '@iqss/dataverse-design-system'
 import { Controller, UseControllerProps, useFieldArray, useFormContext } from 'react-hook-form'
 import { useTranslation } from 'react-i18next'
-import { DynamicFieldsButtons } from '../../../shared/form/DynamicFieldsButtons/DynamicFieldsButtons'
+import { DynamicFieldsButtons } from '@/sections/shared/form/DynamicFieldsButtons/DynamicFieldsButtons'
 import cn from 'classnames'
 import styles from './TopFieldsSection.module.scss'
