+++ src/sections/shared/form/EditCreateCollectionForm/collection-form/top-fields-section/TopFieldsSection.module.scss
